Moves HSTS code to it's own module
